solucionado el problema de envio de mails

+++ src/routes/products.router.js
@@ -4,8 +4,8 @@ import { isAdmin, isPremium } from "../middlewares/auth.middlewares.js";
 import { ErrorMessages } from "../errors/error.enum.js";
 import CustomError from "../errors/CustomError.js";
 import logger from "../winston.js";
-import nodemailer from 'nodemailer';
-import config from "../config.js"
+import EmailService from "../services/email.service.js";
+
 
 const router = Router();
 
@@ -70,21 +70,33 @@ router.post("/",isPremium, async (req, res) => {
 
 router.delete("/:pid", isAdmin || isPremium, async (req, res) => {
   const { pid } = req.params;
+
   try {
     const userId = req.session.user._id;
     const isAdminUser = req.session.user.role === 'ADMIN';
     const product = await productsMongo.findById(pid);
+
+    if (!product) {
+      return res.status(404).json({ message: 'Producto no encontrado' });
+    }
+
     const isOwner = product.owner === userId;
 
     if (isAdminUser || isOwner) {
       const deleteProducts = await productsMongo.deleteProduct(pid);
       logger.info("El producto fue borrado exitosamente");
 
-      if (req.session.user.role === 'premium') {
-       sendDeletedProductEmail(product.title, transporter, res);
-      } else {
-        res.status(200).json({ message: 'Producto borrado' });
-      }
+      const mailOptions = {
+        from: 'marcos.natta@gmail.com',
+        to: product.owner,  
+        subject: 'Producto eliminado',
+        text: `Hola, te informamos que tu producto "${product.title}" ha sido eliminado`,
+      };
+
+      console.log(product.owner)
+      await EmailService.sendEmail(mailOptions);
+
+      res.status(200).json({ message: 'Producto borrado y correo electrónico enviado' });
     } else {
       logger.error("El usuario no tiene permisos para borrar este producto");
       res.status(403).json({ message: "No tienes permisos para borrar este producto" });
@@ -94,34 +106,6 @@ router.delete("/:pid", isAdmin || isPremium, async (req, res) => {
     res.status(500).json({ error: 'Error al intentar borrar el producto', details: error.message });
   }
 });
-
-async function sendDeletedProductEmail(product, res) {
-  try {
-    const transporter = nodemailer.createTransport({
-      service: 'gmail',
-      auth: {
-        user: config.gmail_user,
-        pass: config.gmail_password,
-      },
-    });
-
-    const mailOptions = {
-      from: 'marcos.natta@gmail.com',
-      to: product.owner,
-      subject: 'Producto eliminado',
-      text: `Hola, te informamos que tu producto  ha sido eliminado`,
-    };
-
-    const info = await transporter.sendMail(mailOptions);
-    console.log('Correo electrónico enviado:', info.response);
-    res.status(200).json({ message: 'Producto borrado y correo electrónico enviado' });
-  } catch (error) {
-    console.error('Error al enviar el correo electrónico:', error);
-    res.status(500).json({ error: 'Error al enviar el correo electrónico' });
-  }
-}
-
-
 router.put("/:pid", async (req, res) => {
   const { pid } = req.params;
